Add mapper to support Modal Windows on android
