+++ packages/react-native-session-replay/android/src/main/kotlin/com/datadog/reactnative/sessionreplay/mappers/ReactViewModalMapper.kt
@@ -1,0 +1,10 @@
+package com.datadog.reactnative.sessionreplay.mappers
+
+import ReactViewBackgroundDrawableUtils
+import com.datadog.reactnative.sessionreplay.utils.DrawableUtils
+import com.facebook.react.views.modal.ReactModalHostView
+
+internal class ReactViewModalMapper(
+    private val drawableUtils: DrawableUtils =
+        ReactViewBackgroundDrawableUtils()
+) : DefaultMapper<ReactModalHostView>(drawableUtils)
